+++ neoforge/src/main/java/me/giiena/config/impl/ConfigSyncEvents.java
@@ -1,22 +1,23 @@
 package me.giiena.config.impl;
 
-import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.network.protocol.configuration.ServerConfigurationPacketListener;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.event.OnDatapackSyncEvent;
-import net.neoforged.neoforge.event.entity.player.PlayerEvent;
+import net.neoforged.neoforge.network.event.RegisterConfigurationTasksEvent;
 
 @EventBusSubscriber
 public final class ConfigSyncEvents {
     @SubscribeEvent
-    public static void onPlayerLogin(PlayerEvent.PlayerLoggedInEvent event) {
-        if (event.getEntity() instanceof ServerPlayer player) {
-            ConfigManager.onPlayerLogin(player);
-        }
+    private static void onRegisterConfigTasks(RegisterConfigurationTasksEvent event) {
+        ServerConfigurationPacketListener listener = event.getListener();
+        ConfigManager.onConfigureConnection(listener,
+                event::register,
+                listener::finishCurrentTask);
     }
 
     @SubscribeEvent
-    public static void onServerReload(OnDatapackSyncEvent event) {
+    private static void onServerReload(OnDatapackSyncEvent event) {
         if (event.getPlayer() == null) {
             ConfigManager.onServerReload();
         }